week_10/SQLAgent/requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0